+++ app/Campaigns/routes.js
@@ -6,5 +6,6 @@ Router.get('/', controller.index)
 Router.post('/', request, controller.store)
 Router.get('/:id', controller.show)
 Router.put('/:id', controller.save)
+Router.delete('/:id', controller.remove)
 
 module.exports = Router